feat: remove --experimental-legacy-cli and legacy implementation
