+++ packages/waku/package.json
@@ -30,10 +30,6 @@
       "types": "./dist/middleware/*.d.ts",
       "default": "./dist/middleware/*.js"
     },
-    "./unstable_internals": {
-      "types": "./dist/unstable_internals.d.ts",
-      "default": "./dist/unstable_internals.js"
-    },
     "./unstable_hono": {
       "types": "./dist/unstable_hono.d.ts",
       "default": "./dist/unstable_hono.js"
