--- packages/waku/src/lib/plugins/vite-plugin-rsc-resolve-rsdw.ts
@@ -1,14 +1,0 @@
-import type { Plugin } from 'vite';
-
-export function rscResolveRsdwPlugin(): Plugin {
-  return {
-    name: 'rsc-resolve-rsdw-plugin',
-    enforce: 'pre',
-    async resolveId(id) {
-      if (id === 'react-server-dom-webpack/server.edge') {
-        const resolved = await this.resolve(id);
-        return resolved?.id;
-      }
-    },
-  };
-}
