--- packages/waku/src/middleware/dev-server.ts
@@ -1,1 +1,0 @@
-export { devServer as default } from '../lib/middleware/dev-server.js';
