--- packages/waku/src/middleware/handler.ts
@@ -1,1 +1,0 @@
-export { handler as default } from '../lib/middleware/handler.js';
