--- packages/waku/src/unstable_internals.ts
@@ -1,22 +1,0 @@
-// These exports are for internal use only and subject to change without notice.
-
-export { build } from './lib/builder/build.js';
-
-export { deployAwsLambdaPlugin } from './lib/plugins/vite-plugin-deploy-aws-lambda.js';
-export { deployCloudflarePlugin } from './lib/plugins/vite-plugin-deploy-cloudflare.js';
-export { deployDenoPlugin } from './lib/plugins/vite-plugin-deploy-deno.js';
-export { deployNetlifyPlugin } from './lib/plugins/vite-plugin-deploy-netlify.js';
-export { deployPartykitPlugin } from './lib/plugins/vite-plugin-deploy-partykit.js';
-export { deployVercelPlugin } from './lib/plugins/vite-plugin-deploy-vercel.js';
-export { devCommonJsPlugin } from './lib/plugins/vite-plugin-dev-commonjs.js';
-export { nonjsResolvePlugin } from './lib/plugins/vite-plugin-nonjs-resolve.js';
-export { rscAnalyzePlugin } from './lib/plugins/vite-plugin-rsc-analyze.js';
-export { rscDelegatePlugin } from './lib/plugins/vite-plugin-rsc-delegate.js';
-export { rscEntriesPlugin } from './lib/plugins/vite-plugin-rsc-entries.js';
-export { rscEnvPlugin } from './lib/plugins/vite-plugin-rsc-env.js';
-export { rscHmrPlugin } from './lib/plugins/vite-plugin-rsc-hmr.js';
-export { rscIndexPlugin } from './lib/plugins/vite-plugin-rsc-index.js';
-export { rscManagedPlugin } from './lib/plugins/vite-plugin-rsc-managed.js';
-export { rscPrivatePlugin } from './lib/plugins/vite-plugin-rsc-private.js';
-export { rscRsdwPlugin } from './lib/plugins/vite-plugin-rsc-rsdw.js';
-export { rscTransformPlugin } from './lib/plugins/vite-plugin-rsc-transform.js';
