--- packages/waku/tests/fixtures/plugin-rsc-analyze/client.ts
@@ -1,2 +1,0 @@
-'use client';
-console.log('This is client side code!');
