--- packages/waku/tests/fixtures/plugin-rsc-analyze/import-client.ts
@@ -1,1 +1,0 @@
-import './client.js';
