--- packages/waku/tests/fixtures/plugin-rsc-analyze/import-server.ts
@@ -1,1 +1,0 @@
-import './server.js';
