--- packages/waku/tests/fixtures/plugin-rsc-analyze/server.ts
@@ -1,2 +1,0 @@
-'use server';
-console.log('This is server side code');
